Move JobDrivers into their own subfolder and adjust namespace
